Capitalize "P" in "Psalms" in WCF 21.5
This is capitalized in the recently published critical edition[1]
and there are no variants noted.

Although capitalization did not carry the same meaning or have the same
consistency as today, it's reasonable to conclude that this paragraph is
referring to the canonical book of Psalms based on contemporary writings
and the practice of the Scottish church.

This lack of capitalization in modern reproductions has been appealed
to by proponents of uninspired songs in worship to deny the Westminster
Assembly spoke of the Psalter. This capitalization change is meaningful

[1]: John Bower, “Chap. XXI. Of Religious Worship, and the Sabbath Day,”
in The Confession of Faith: a Critical Text and Introduction
(Grand Rapids, MI, MI: Reformation Heritage Books, 2020), pp. 221-221.